fix: auth types and Discord ID fetching
`next-auth` changed what `DiscordProvider` returns in the token subject between
beta versions and it gives us back a UUID, even though we should be expecting
a Discord snowflake user ID. There is no other way to fetch the snowflake ID
other than that jank profile picture URL splitting, but it works, so be it.

Finally, extend the types from `next-auth` instead of defining our own, and use
the `access_token` naming for consistency.

Thanks to the following resources for the contents of this commit:
- https://github.com/nextauthjs/next-auth/discussions/9120#discussioncomment-7544307
- https://github.com/nextauthjs/next-auth/issues/7122#issuecomment-2591137527